refactor(schemas): move nextlyMeta to schemas/nextly-meta/
Per-dialect Drizzle table definitions for nextlyMeta (the runtime key/value
flags table) now live in schemas/nextly-meta/{postgres,mysql,sqlite}.ts
with a dialect-aware barrel exposing nextlyMetaTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports keep compiling.
